### 2024-04-16 #6635 @editorialbot
[PRE REVIEW]: PowerAPI: A Python framework for building software-defined power meters
**Submitting author:** @roda82 (<a href="http://orcid.org/0000-0002-5317-2610">Daniel Romero Acero</a>)
**Repository:** https://github.com/powerapi-ng/powerapi
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** V2.6.0
