+++ paper.bib
@@ -25,16 +25,6 @@
   url = {https://github.com/powerapi-ng/smartwatts-formula}
 }
 
-
-@misc{hwpc-sensor,
-  author = {G. Fieni and R. Rouvoy},
-  title = {Hardware Performance Counters (HwPC) Sensor},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/powerapi-ng/hwpc-sensor}
-}
-
 @inproceedings{fieni:2020,
   TITLE = {{SmartWatts: Self-Calibrating Software-Defined Power Meter for Containers}},
   AUTHOR = {Fieni, Guillaume and Rouvoy, Romain and Seinturier, Lionel},
@@ -48,6 +38,7 @@
   PDF = {https://inria.hal.science/hal-02470128/file/smartwatts-paper.pdf},
   HAL_ID = {hal-02470128},
   HAL_VERSION = {v1},
+  DOI = {10.1109/CCGrid49817.2020.00-45},
 }
 
 @article{pedregosa:2011,
@@ -57,7 +48,8 @@
   volume={12},
   number={Oct},
   pages={2825--2830},
-  year={2011}
+  year={2011},
+  doi={10.5555/1953048.2078195},
 }
 
 @manual{intel:2023,
@@ -80,7 +72,8 @@ booktitle = {Proceedings of the 3rd International Joint Conference on Artificial
 pages = {235–245},
 numpages = {11},
 location = {Stanford, USA},
-series = {IJCAI'73}
+series = {IJCAI'73},
+doi = {10.5555/1624775.1624804},
 }
 
 @book{agha:1986,
@@ -104,6 +97,7 @@ address = {Cambridge, MA, USA}
   PDF = {https://inria.hal.science/hal-03173410/file/selfwatts-paper.pdf},
   HAL_ID = {hal-03173410},
   HAL_VERSION = {v1},
+  DOI = {10.1109/ccgrid51090.2021.00042},
 }
 
 @article{colmant:2018,
@@ -139,6 +133,7 @@ address = {Cambridge, MA, USA}
   PDF = {https://inria.hal.science/hal-01439889/file/colmant-ccgrid17-camera-ready.pdf},
   HAL_ID = {hal-01439889},
   HAL_VERSION = {v1},
+  DOI = {10.1109/ccgrid.2017.27},
 }
 
 @inproceedings{lebeane:2015,
