+++ paper.bib
@@ -49,7 +49,6 @@
   number={Oct},
   pages={2825--2830},
   year={2011},
-  doi={10.5555/1953048.2078195},
 }
 
 @manual{intel:2023,
@@ -73,7 +72,6 @@ pages = {235–245},
 numpages = {11},
 location = {Stanford, USA},
 series = {IJCAI'73},
-doi = {10.5555/1624775.1624804},
 }
 
 @book{agha:1986,
