### 2024-04-17 #6635 @roda82
Hello, we changed the title to `PowerAPI: A Python framework for building software-defined power meters`. Can you please update the metadata related to our submission?
