### 2024-04-23 e28b364
fix: Rename parameters names in extract_file_names action, Remove code related to influxdb 1.8
