### 2024-04-23 #6635 @danielskatz
Thanks @varshaprasad96 - the information is in https://joss.readthedocs.io/en/latest/reviewer_guidelines.html  At this point, I will add you in the system as a reviewer, but we won't start the review until we get at least one more reviewer.
