### 2024-04-23 #6670 @editorialbot
[REVIEW]: PowerAPI: A toolkit for estimating power consumption of software
**Submitting author:** @roda82 (<a href="http://orcid.org/0000-0002-5317-2610">Daniel Romero Acero</a>)
**Repository:** https://github.com/powerapi-ng/powerapi
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v2.7.0
