### 2024-04-25 22063a4
build(deps): bump python from `399ca82` to `2be8dad`
Bumps python from `399ca82` to `2be8dad`.

---
updated-dependencies:
- dependency-name: python
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
