+++ deploy/docker/README.md
@@ -1,0 +1,498 @@
+# PowerAPI Docker Toolchain
+
+This folder contains several resources to test and configure PowerAPI on Docker based hosting platform using docker-compose.
+
+
+**NOTE:** 
+* This toolchain is meant for running quick tests 
+  of the PowerAPI energy metering solution
+  and is not suitable for detailed analysis.
+* For simplicity sake all components are deployed on a single host. 
+  While it makes it much easier to quickly experiment with PowerAPI monitoring,
+  it should not be used when running actual tests. You should only run the sensor
+  on your target systems and all other components on an independent system
+  used only for monitoring.
+* This toolchain eventually can be used to test, select and build your own deployment
+  based on current settings and diagrams available in this folder.
+
+## Prerequisites
+
+* a linux system (tested on ubuntu 20.04 and 22.04)
+* docker
+* docker-compose
+* internet access for docker registry
+
+## Quick Start
+
+No build step is needed, the default settings based on [mongodb/prometheus](#mongodb_prometheus) configuration can be launched "Out  of the box" with command below:
+
+```
+docker-compose up
+```
+
+Always stop properly components to unmount associated resources:
+
+```
+docker-compose down
+```
+
+For more accurate measurements, you should configure the formula according to the cpu on you host. This can be done by configuring CPU profile in _./formula_ folder configuration files.
+Look at [the offical powerapi documentation](https://powerapi-ng.github.io/howto_monitor_process/deploy_formula.html) to see which value to use for `cpu-ratio-base`,`cpu-ratio-min` and `cpu-ratio-max`.
+
+## Configuration
+
+Based on docker-compose **profiles** and **.env** file, several configurations are available based on [Sources/Destinations](https://powerapi.org/reference/database/sources_destinations/) matrix:
+
+| Source  | Destination |                             |
+|---------|-------------|-----------------------------|
+| socket  | csv         | [view](#socket_csv)         |
+| socket  | filedb      | [view](#socket_filedb)      |
+| socket  | influxdb2   | [view](#socket_influxdb2)   |
+| socket  | mongodb     | [view](#socket_mongodb)     |
+| socket  | prometheus  | [view](#socket_prometheus)  |
+| mongodb | csv         | [view](#mongodb_csv)        |
+| mongodb | filedb      | [view](#mongodb_filedb)     |
+| mongodb | influxdb2   | [view](#mongodb_influxdb2)  |
+| mongodb | mongodb     | [view](#mongodb_mongodb)    |
+| mongodb | prometheus  | [view](#mongodb_prometheus) |
+
+Source and Destination selection can be done editing **.env** file as follow:
+```properties
+# Source selection
+# Available options: mongodb, socket
+POWERAPI_SOURCE=mongodb
+
+# Destination selection
+# Available options: influxdb2, prometheus, mongodb, csv, filedb
+POWERAPI_DESTINATION=prometheus
+```
+
+Currently, only [HWPC Sensor](https://powerapi.org/reference/sensors/hwpc-sensor/) and [SmartWatts Formula](https://powerapi.org/reference/formulas/smartwatts/) are supported and can be modified with properties below:
+```properties
+# Sensor selection
+# Available options: hwpc
+SENSOR_TYPE=hwpc
+SENSOR_IMAGE=powerapi/${SENSOR_TYPE}-sensor:1.3.0
+
+# Formula selection
[…]
