+++ deploy/docker/grafana/datasource-prometheus.yml
@@ -1,0 +1,10 @@
+apiVersion: 1
+datasources:
+- orgId: 1
+  version: 1
+  name: prometheus-powerapi
+  type: prometheus
+  access: proxy
+  url: http://prometheus:9090
+  isDefault: true
+  editable: true
