+++ deploy/docker/docker-compose.yaml
@@ -1,0 +1,171 @@
+version: '3.8'
+
+services:
+  ############################################
+  #          SOURCES & DESTINATIONS          #
+  ############################################
+  # MongoDB
+  mongodb:
+    container_name: mongodb
+    image: ${MONGO_IMAGE}
+    networks:
+      - powerapi-network
+    profiles:
+      - mongodb
+  
+  # InluxDB2
+  influxdb2:
+    container_name: influxdb2
+    image: ${INFLUXDB_IMAGE}
+    environment:
+      DOCKER_INFLUXDB_INIT_MODE: "setup"
+      DOCKER_INFLUXDB_INIT_USERNAME: "powerapi"
+      DOCKER_INFLUXDB_INIT_PASSWORD: "powerapi"
+      DOCKER_INFLUXDB_INIT_BUCKET: "powerapi"
+      DOCKER_INFLUXDB_INIT_ORG: "powerapi"      
+      DOCKER_INFLUXDB_INIT_ADMIN_TOKEN: "powerapi"
+    ports: 
+      - "8086:8086"
+    volumes: 
+      - ${PWD}/influxdb:/var/lib/influxdb 
+    networks:
+      - powerapi-network
+    profiles:
+      - influxdb2
+
+  # Prometheus
+  prometheus:
+    container_name: prometheus
+    image: ${PROMETHEUS_IMAGE}
+    ports: 
+     - "9090:9090"
+    volumes: 
+      - ${PWD}/prometheus:/opt/bitnami/prometheus/data # Folder must be writable for UID 1001
+      - ${PWD}/prometheus/prometheus.yml:/etc/prometheus/prometheus.yml
+    networks:
+      - powerapi-network
+    profiles:
+      - prometheus
+
+  # CSV (Default is a dummy image, can be used to add treatments on CSV files)
+  csv:
+    container_name: csv
+    image: ${CSV_IMAGE}
+    volumes:
+      - ${PWD}/csv:/tmp/csv
+    profiles:
+      - csv
+
+  # FileDB (Default is a dummy image, can be used to add treatments on data files)
+  filedb:
+    container_name: filedb
+    image: ${FILEDB_IMAGE}
+    volumes:
+      - ${PWD}/filedb:/tmp/filedb
+    profiles:
+      - filedb
+
+  # Socket (Default is a dummy image, can be used to add treatments on formula socket)
+  socket:
+    container_name: socket
+    image: ${SOCKET_IMAGE}
+    profiles:
+      - socket
+
+  ############################################
+  #                POWER API                 #
+  ############################################
+  # PowerAPI Sensor
+  sensor:
+    container_name: sensor
+    image: ${SENSOR_IMAGE}
+    privileged: true
+    command:
+      - "--config-file"
+      - "/etc/sensor.json"
+    volumes:
+      - ${PWD}/sensor/${SENSOR_TYPE}-${POWERAPI_SOURCE}.json:/etc/sensor.json
+      - ${PWD}/csv:/tmp/csv
+      - type: bind
+        source: /proc
+        target: /proc
+      - type: bind
+        source: /sys
+        target: /sys
+      - type: bind
+        source: /var/lib/docker/containers
+        target: /var/lib/docker/containers
+    depends_on:
+      - ${POWERAPI_SOURCE}
+    networks:
+      - powerapi-network
+    restart: unless-stopped
+    
+  # PowerAPI Formula
+  formula:
+    container_name: formula
+    image: ${FORMULA_IMAGE}
+    command:
+      - "--config-file"
+      - "/etc/formula.json"
+    volumes:
[…]
