### 2024-04-29 7209c96
fix: Correct parsing of environment variable by using longest argument suffix
