### 2024-04-29 9f5a60a
fix: Adapt test to changes on environment variables parsing
