### 2024-05-12 #6670 @EpicStep
@roda82 it is possible to mark 'master' as default instead of 'gh-pages' branch at https://github.com/powerapi-ng/powerapi-ng.github.io? Or add some description that there is a 'master' branch and it is necessary to contribute into it?
