### 2024-05-12 #6670 @EpicStep
@roda82 I see `docs/` directory in https://github.com/powerapi-ng/powerapi. It seemed outdated to me. Should we remove it?
