### 2024-05-14 5235070
refactor(pusher): Fix pylint `useless-return` warning
