+++ .github/workflows/build.yml
@@ -44,7 +44,7 @@ jobs:
 
     - name: Upload coverage reports to Codecov
       if: ${{ matrix.python-version }} == "3.10"
-      uses: codecov/codecov-action@5ecb98a3c6b747ed38dc09f787459979aebb39be # v4.3.1
+      uses: codecov/codecov-action@6d798873df2b1b8e5846dba6fb86631229fbcb17 # v4.4.0
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         file: coverage.xml
