+++ docs/old/dev.rst
@@ -1,145 +1,0 @@
-.. Concept page
-
-PowerAPI - Guide développeur
-******************************
-
-Smartwatts est implémenté autour d'un cadriciel d'acteur, les acteurs communiquent grâce à la bibliothèque `pyzmq <https://pyzmq.readthedocs.io/en/latest/>`_ qui est une implémentation python de la librairie `ZeroMQ <http://zeromq.org/>`_. Ce guide défini les principaux concepts dont la compréhension est nécessaire pour utiliser ces acteurs.
-
-Définition d'un acteur
-======================
-
-Les acteurs utilisés dans powerapi reposent tous sur la même classe abstraite :class:`powerapi.Actor <powerapi.actor.actor.Actor>` qui hérite elle même de la classe `multiprocessing.Process <https://docs.python.org/3/library/multiprocessing.html>`_. Cette classe définie un acteur comme une entité autonome adaptant son comportement en fonction des messages qu'elle reçoit.
-
-Concrètement, un acteur est un processus système. Ce processus possède un comportement prédéfini (son comportement métier), un état (un ensemble de valeur stocké par l'acteur tout au long de son exécution) et une interface de communication avec les autres acteurs. Il est capable d'envoyer/recevoir des messages à/depuis d'autres processus. En fonction des messages reçus, l'acteur peut :
-
-- modifier son état
-- produire des effets de bord (contacter/créer d'autres acteurs, intéragir avec le système, ...)
-- changer son comportement métier.
-
-Interface client/serveur
-========================
-
-Une interface client/serveur expose les fonctions permettant de manipuler les acteurs (communication, configuration, ...)
-
-l'interface serveur regroupe l'ensemble des fonctions internes à l'acteur (communication avec d'autres processus, modification de l'état de l'acteur, ...). Ces fonctions sont exécutées dans le processus de l'acteur. Elles regroupent : la configuration de l'acteur côté serveur, la réception et le traitement des messages.
-
-l'interface client regroupe l'ensemble des fonctions permettant de lancer et configurer un acteur ainsi que les fonctions permettant à un autre processus de communiquer avec l'acteur. Elles sont exécutées dans les processus souhaitant communiquer avec l'acteur. Ces fonctions servent à initialiser/configurer les variables de communication avec l'acteur et lui envoyer des messages.
-
-- le code de la classe :class:`powerapi.Actor <powerapi.actor.actor.Actor>` regroupe à la fois les fonctions de l'interface client comme celle de l'interface serveur. La méthode :func:`Actor.start` de l'acteur va dupliquer (fork) le code de la classe acteur et créer le processus système. Les variables permettant d'utiliser l'interface serveur seront alors initialisées dans le nouveau processus. Ainsi, les fonctions de l'interface serveur ne peuvent être utilisées dans le processus client.
-- La méthode :func:`Actor.connect() <powerapi.actor.actor.Actor.connect>` permet d'initialiser les variables permettant d'utiliser l'interface client dans un processus souhaitant communiquer avec le serveur.
-
-Composantes d'un acteur
-=======================
-
-La classe :class:`powerapi.Actor <powerapi.actor.actor.Actor>` encapsule plusieurs composants. Chacun pouvant être enrichie pour personnaliser le comportement de l'acteur.
-
-État
------
-
-L'état de l'acteur est un ensemble de valeurs utilisées par l'acteur lorsqu'il traite un message. Cet état est implémenté par la classe :class:`powerapi.BasicState <powerapi.actor.state.BasicState>`.
-
-Cette classe contient les valeurs suivantes :
-
-- l'interface de communication
[…]
