### 2024-05-16 5f7ce2b
refactor: Remove unnecessary files and directory on source
