### 2024-05-16 #6670 @roda82
> @roda82 it is possible to mark 'master' as default instead of 'gh-pages' branch at https://github.com/powerapi-ng/powerapi-ng.github.io? Or add some description that there is a 'master' branch and it is necessary to contribute into it?

Thank you for your feedback. I just marked `master` as default. However, we are not really waiting for contributions on the repository related to the website.
