### 2024-05-16 #6670 @roda82
> @roda82 I see `docs/` directory in https://github.com/powerapi-ng/powerapi. It seemed outdated to me. Should we remove it?

I just removed some unnecessary files. We cannot remove `docs` as it is used by `sphinx` to generate the API documentation by using docstring.
