### 2024-05-20 #6670 @danielskatz
👋 @varshaprasad96 it looks like @roda82 has addressed your two open issues. Can these now be closed? Do they let you continue your review?

Thanks!
