+++ paper.md
@@ -31,14 +31,14 @@ bibliography: paper.bib
 
 # Summary
 
-Software that we use daily for accessing digital services from connected devices has a negative impact on the environment as it consumes energy. These digital services, hosted by physical machines around the world, also contribute to the planet pollution. Unfortunately, providers of these online services mostly focus on hardware efficiency to reduce the environmental impact without considering the software they host. For this reason, we propose [`PowerAPI`](https://powerapi.org/) [@powerapi-core], a software-defined solution delivering real-time estimations of software power consumption to spot opportunities to reduce it and therefore limiting their impact on the planet beyond hardware improvements.
+Software that we use daily for accessing digital services from connected devices has a negative impact on the environment as it consumes energy. These digital services, hosted by physical machines around the world, also contribute to planetary pollution. Unfortunately, providers of these online services mostly focus on hardware efficiency to reduce the environmental impact without considering the software they host. For this reason, we propose [`PowerAPI`](https://powerapi.org/) [@powerapi-core], a software-defined solution that delivers real-time estimations of software power consumption to spot opportunities to reduce it and therefore to limit their impact on the planet beyond hardware improvements.
 
 
 # Statement of need
 
-State-of-the-art power meters only deliver a global consumption of machines hosting the software. To deal with this limitation, several software-defined power meters has been proposed. However, they are statically designed and coupled to a specific processor family or they are mainly based on power models that need a lot of data and time to be trained to predict the power consumption accurately [@colmant:2018] [@lebeane:2015]. Power models designed in this way are only suitable for similar environmental conditions (i.e., memory usage, CPU usage, kinds of running applications...) when data for the training was collected. If these conditions change, power models become deprecated and have to be trained again, which make them unsuitable for production environments of Cloud providers.
+State-of-the-art power meters only measure the global consumption of power of the machines hosting the software. To deal with this limitation, several software-defined power meters have been proposed. However, they are statically designed and coupled to a specific processor family or they are mainly based on power models that need a lot of data and time to be trained to predict power consumption accurately [@colmant:2018; @lebeane:2015]. Power models designed in this way are only suitable for environmental conditions (i.e., memory usage, CPU usage, kinds of running applications) similar to those when the data for the training was collected. If these conditions change, power models become deprecated and have to be trained again, which make them unsuitable for production environments of Cloud providers.
 
-To deal with this constraint, we developed [`PowerAPI`](https://powerapi.org/) [@powerapi-core], a software toolkit for assembling software power meters enabling developers/IT administrators to monitor power consumption of software that they write/deploy. Software-defined power meters created with `PowerAPI` enable power consumption estimations at different granularity levels: process, thread, container, virtual machine, etc. Furthermore, power models used by `PowerAPI` are continuously self-calibrated to consider current conditions execution of the machine hosting the software.  
[…]
