+++ paper.bib
@@ -26,10 +26,10 @@
 }
 
 @inproceedings{fieni:2020,
-  TITLE = {{SmartWatts: Self-Calibrating Software-Defined Power Meter for Containers}},
+  TITLE = {Smart{W}atts: Self-Calibrating Software-Defined Power Meter for Containers},
   AUTHOR = {Fieni, Guillaume and Rouvoy, Romain and Seinturier, Lionel},
   URL = {https://inria.hal.science/hal-02470128},
-  BOOKTITLE = {{CCGRID 2020 - 20th IEEE/ACM International Symposium on Cluster, Cloud and Internet Computing}},
+  BOOKTITLE = {CCGRID 2020 - 20th IEEE/ACM International Symposium on Cluster, Cloud and Internet Computing},
   ADDRESS = {Melbourne, Australia},
   YEAR = {2020},
   MONTH = May,
@@ -42,7 +42,7 @@
 }
 
 @article{pedregosa:2011,
-  title={Scikit-learn: Machine learning in Python},
+  title={Scikit-learn: Machine learning in {P}ython},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
   journal={Journal of machine learning research},
   volume={12},
@@ -99,10 +99,10 @@ address = {Cambridge, MA, USA}
 }
 
 @article{colmant:2018,
-  TITLE = {{The Next 700 CPU Power Models}},
+  TITLE = {The Next 700 {CPU} Power Models},
   AUTHOR = {Colmant, Maxime and Rouvoy, Romain and Kurpicz, Mascha and Sobe, Anita and Felber, Pascal and Seinturier, Lionel},
   URL = {https://inria.hal.science/hal-01827132},
-  JOURNAL = {{Journal of Systems and Software}},
+  JOURNAL = {Journal of Systems and Software},
   PUBLISHER = {{Elsevier}},
   VOLUME = {144},
   PAGES = {382-396},
@@ -116,10 +116,10 @@ address = {Cambridge, MA, USA}
 }
 
 @inproceedings{colmant:2017,
-  TITLE = {{WattsKit: Software-Defined Power Monitoring of Distributed Systems}},
+  TITLE = {Watts{K}it: Software-Defined Power Monitoring of Distributed Systems},
   AUTHOR = {Colmant, Maxime and Felber, Pascal and Rouvoy, Romain and Seinturier, Lionel},
   URL = {https://inria.hal.science/hal-01439889},
-  BOOKTITLE = {{17th IEEE/ACM International Symposium on Cluster, Cloud and Grid Computing (CCGrid)}},
+  BOOKTITLE = {17th IEEE/ACM International Symposium on Cluster, Cloud and Grid Computing ({CCGrid})},
   ADDRESS = {Madrid, Spain},
   EDITOR = {Franck Capello and Geoffrey Fox and Javier Garcia-Blas},
   PUBLISHER = {{IEEE}},
