### 2024-05-22 #6670 @danielskatz
@roda82 - I've proofread the paper, and suggest changes in the paper/bib in https://github.com/powerapi-ng/powerapi/pull/341. Please merge this, or let me know what you disagree with, then we can continue the acceptance process.
