### 2024-05-24 956ef49
fix: Change filtering by tags on PowerReport. Now, if tags is empty all metadata is kept
