### 2024-05-24 3c3ea38
fix: Remove useless tags initialisation in save method of prometheus_db
