### 2024-05-29 48b49a0
docs: Add CITATION.cff file
