### 2024-05-29 78c3397
feat(database/socket): Rework socket database code
