+++ src/powerapi/database/socket_db.py
@@ -27,99 +27,151 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-import asyncio
-from typing import Type, List
-import json
+import logging
+from json import JSONDecoder, JSONDecodeError
+from queue import SimpleQueue, Empty
+from socketserver import ThreadingMixIn, TCPServer, StreamRequestHandler
+from threading import Thread
+from typing import Type, Iterator
 
-from powerapi.utils import JsonStream
+from powerapi.database.base_db import IterDB, BaseDB
 from powerapi.report import Report
-from .base_db import IterDB, BaseDB, DBError
+
+
+class ThreadedTCPServer(ThreadingMixIn, TCPServer):
+    """
+    TCP Server implementation.
+    Each client connected will be served by a separate thread.
+    """
+    daemon_threads = True
+    allow_reuse_address = True
+
+    def __init__(self, server_address, request_handler_class, received_data_queue: SimpleQueue):
+        """
+        :param server_address: The address to listen on
+        :param request_handler_class: The request handler class to use when receiving requests
+        :param received_data_queue: The data queue to store the received data
+        """
+        super().__init__(server_address, request_handler_class)
+        self.received_data_queue = received_data_queue
+
+
+class JsonRequestHandler(StreamRequestHandler):
+    """
+    Request handler that handles JSON documents received from the client.
+    """
+    server: ThreadedTCPServer
+
+    @staticmethod
+    def parse_json_documents(data: str) -> Iterator[dict]:
+        """
+        Try to parse json document(s) from the given string.
+        This function tolerates truncated and malformed json documents.
+        :param data: The raw data to decode
+        :return: Iterator over parsed json documents
+        """
+        decoder = JSONDecoder()
+        idx = 0
+        while idx < len(data):
+            try:
+                obj, end_idx = decoder.raw_decode(data, idx)
+                yield obj
+                idx += end_idx
+
+            # Search and try to parse the remaining document(s)
+            except JSONDecodeError as e:
+                idx = data.find('{', e.pos)
+                if idx == -1:
+                    break
+
+    def handle(self):
+        """
+        Handle incoming connections.
+        The received data is parsed and the result(s) stored in the data queue for further processing.
+        It is expected for the data to be in json format (utf-8 charset) and newline terminated.
+        """
+        caddr = '{}:{}'.format(*self.client_address)
+        logging.info('New incoming connection from %s', caddr)
+
+        while True:
+            try:
+                data = self.rfile.readline()
+                if not data:
+                    break
+
+                for obj in self.parse_json_documents(data.decode('utf-8')):
+                    self.server.received_data_queue.put(obj)
+
+            except ValueError as e:
+                logging.warning('[%s] Received malformed data: %s', caddr, e)
+                continue
+            except OSError as e:
+                logging.error('[%s] Caught OSError while handling request: %s', caddr, e)
+                break
+            except KeyboardInterrupt:
+                break
+
+        logging.info('Connection from %s closed', caddr)
 
 
 class IterSocketDB(IterDB):
     """
-    iterator connected to a socket that receive report from a sensor
+    SocketDB iterator that returns the received data.
     """
 
-    def __init__(self, report_type, stream_mode, queue):
-        """
-        """
-        IterDB.__init__(self, None, report_type, stream_mode)
-
-        self.queue = queue
-
-    def __aiter__(self):
+    def __iter__(self):
         return self
 
-    async def __anext__(self):
+    def __next__(self):
         try:
[…]
