### 2024-05-30 ccd1f87
test: Add tests for PowerReport related to tags management and generation of influxdb and prometheus documents
