+++ tests/unit/report/test_power_report.py
@@ -35,6 +35,72 @@ from datetime import datetime
 from tests.utils.report.power import gen_json_power_report
 
 
+def get_expected_metadata_power_report_without_tag_list(report: PowerReport) -> dict:
+    """
+    Generates the expected metadata dict for the report with the gen_tag function
+    :param report The report to extract the metadata
+    """
+    metadata = report.metadata
+    metadata['sensor'] = report.sensor
+    metadata['target'] = report.target
+
+    return metadata
+
+
+def get_expected_metadata_power_report_with_tags(report: PowerReport, tags: list) -> dict:
+    """
+    Generates the expected metadata dict for the report with the gen_tag function
+    :param report: The report to extract the metadata
+    :param tags: The tags to be kept
+    """
+    metadata = {'sensor': report.sensor, 'target': report.target}
+
+    for tag in tags:
+        metadata[tag] = report.metadata[tag]
+
+    return metadata
+
+
+def get_expected_influxdb_document(report: PowerReport, tags: list) -> dict:
+    """
+    Generates a dictionary that represents the expected influxdb document for a given dictionary
+    :param report: The report for generating the document
+    :param tags: The tags to be kept
+    """
+    return {
+        'measurement': 'power_consumption',
+        'tags': get_expected_metadata_power_report_with_tags(report, tags) if tags else
+        get_expected_metadata_power_report_without_tag_list(report),
+        'time': str(report.timestamp),
+        'fields': {
+            'power': report.power
+        }
+    }
+
+
+def get_expected_prometheus_document(report: PowerReport, tags: list) -> dict:
+    """
+    Generates a dictionary that represents the expected prometheus document for a given dictionary
+    :param report: The report for generating the document
+    :param tags: The tags to be kept
+    """
+    return {
+        'tags': get_expected_metadata_power_report_with_tags(report, tags) if tags else
+        get_expected_metadata_power_report_without_tag_list(report),
+        'time': int(report.timestamp.timestamp()),
+        'value': report.power
+    }
+
+
+def check_report_metadata(original_metadata: dict, report: PowerReport):
+    """
+    Check that the metadata of a report didn't change
+    :param original_metadata: Orignal's report metadata
+    :param report: Report for checking metadata
+    """
+    assert report.metadata == original_metadata
+
+
 ########
 # JSON #
 ########
@@ -155,3 +221,325 @@ def test_create_report_from_csv_with_metadata():
                  ]
     report = PowerReport.from_csv_lines(csv_lines)
     assert report.metadata["tag"] == 1
+
+
+def test_gen_tag_keep_all_the_report_metadata_without_tags_list_and_empty_metadata(power_report_without_metadata):
+    tags = []
+    original_metadata = power_report_without_metadata.metadata
+    expected_metadata = get_expected_metadata_power_report_without_tag_list(power_report_without_metadata)
+
+    metadata = power_report_without_metadata.gen_tag(tags)
+
+    assert metadata == expected_metadata
+    check_report_metadata(original_metadata, power_report_without_metadata)
+
+
+def test_gen_tag_keep_all_the_report_metadata_without_tag_list_and_empty_metadata(power_report_without_metadata):
+    tags = None
+    original_metadata = power_report_without_metadata.metadata
+    expected_metadata = get_expected_metadata_power_report_without_tag_list(power_report_without_metadata)
+
[…]
