### 2024-06-03 #6670 @danielskatz
👋 @roda82 - note that we're close to done, and just need a few more things from you, as [above](https://github.com/openjournals/joss-reviews/issues/6670#issuecomment-2124897106). (And I apologize for somehow tagging the wrong person there, so you may not have seen this.)
