### 2024-06-03 9b1450d
feat(config): Add `host` parameter to config the socket db listen address
