### 2024-06-03 c863617
refactor(report): Clean `_extract_timestamp` method code
