### 2024-06-03 b477755
build: Bump PowerAPI package version to 2.7.0
