### 2024-06-04 #6670 @roda82
@danielskatz Thank you for the feedback. I just did what you asked me to do. The DOI is `10.5281/zenodo.11453194` and the PowerAPI version `2.7.0`. I tried to update them without success.  Authors' ORCID are already added.  Please notice that it is also required to change the article title to `PowerAPI: A Python framework for building software-defined power meters`.
