### 2024-06-04 #6670 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
