+++ paper.bib
@@ -84,10 +84,10 @@ address = {Cambridge, MA, USA}
 }
 
 @inproceedings{fieni:2021,
-  TITLE = {{SELFWATTS: On-the-fly Selection of Performance Events to Optimize Software-defined Power Meters}},
+  TITLE = {SELFWATTS: On-the-fly Selection of Performance Events to Optimize Software-defined Power Meters},
   AUTHOR = {Fieni, Guillaume and Rouvoy, Romain and Seinturier, Lionel},
   URL = {https://inria.hal.science/hal-03173410},
-  BOOKTITLE = {{CCGRID 2021 - 21th IEEE/ACM International Symposium on Cluster, Cloud and Internet Computing}},
+  BOOKTITLE = {CCGRID 2021 - 21th IEEE/ACM International Symposium on Cluster, Cloud and Internet Computing},
   ADDRESS = {Melbourne, Australia},
   YEAR = {2021},
   MONTH = May,
