### 2024-06-04 #6670 @danielskatz
@roda82 - sorry, I see one more change needed - please merge https://github.com/powerapi-ng/powerapi/pull/347

Also, please confirm that the proof is fine, other than these case issues in a bib entry.
