### 2024-06-04 #6670 @danielskatz
Congratulations to @roda82 (Daniel Romero Acero) and co-authors on your publication!!

And thanks to @varshaprasad96 and @EpicStep for reviewing!
JOSS depends on volunteers and we couldn't do this without you.
